added login component; moved login link from footer to navbar. Also changed server-render-test to: 4D Nucleome for title
